popup refactor (delete DelPopup.svelte, create popup.ts), doc DelAutocomplete
